hoàn thành crud - jwt - authorize

+++ Services/RoleService.cs
@@ -1,4 +1,5 @@
-﻿using AuthorizationAPI.Models;
+﻿using AuthorizationAPI.DTOs.Request;
+using AuthorizationAPI.Models;
 using AuthorizationAPI.Repositories;
 using AuthorizationAPI.Repositories.IRepo;
 using System.Collections.Generic;
@@ -15,29 +16,81 @@ namespace AuthorizationAPI.Services
             _roleRepository = roleRepository;
         }
 
-        public async Task<IEnumerable<Role>> GetAllRolesAsync()
+        public async Task<ApiResponse<ICollection<Role>>> GetAllRolesAsync()
         {
-            return await _roleRepository.GetAllRolesAsync();
+            var roles = await _roleRepository.GetAllRolesAsync();
+            return new ApiResponse<ICollection<Role>>("success", "Lấy danh sách phân quyền thành công", roles);
         }
 
-        public async Task<Role?> GetRoleByIdAsync(int id)
+        public async Task<ApiResponse<Role>> GetRoleByIdAsync(int id)
         {
-            return await _roleRepository.GetRoleByIdAsync(id);
+            var role = await _roleRepository.GetRoleByIdAsync(id);
+
+            if (role == null)
+            {
+                return new ApiResponse<Role>("error", "Không tìm thấy quyền", null);
+            }
+            return new ApiResponse<Role>("success", "Tìm thấy quyền", role); ;
         }
 
-        public async Task AddRoleAsync(Role role)
+        public async Task<ApiResponse<Role>> AddRoleAsync(RoleRequest request)
         {
-            await _roleRepository.AddRoleAsync(role);
+            try
+            {
+                var newRole = new Role
+                {
+                    RoleName = request.Name,
+                };
+                var role = await _roleRepository.AddRoleAsync(newRole);
+                return new ApiResponse<Role>("success", "Thêm quyền thành công", role);
+            }
+            catch (Exception ex)
+            {
+                return new ApiResponse<Role>("error", "Lỗi thêm quyền", null);
+            }
+            
         }
 
-        public async Task UpdateRoleAsync(Role role)
+        public async Task<ApiResponse<Role>> UpdateRoleAsync(int id, RoleRequest request)
         {
-            await _roleRepository.UpdateRoleAsync(role);
+            try
+            {
+                var existRole = await _roleRepository.GetRoleByIdAsync(id);
+                if (existRole == null)
+                {
+                    return new ApiResponse<Role>("error", "Không tìm thấy người dùng", null);
+                }
+
+                existRole.RoleName = request.Name;
+                var updateRole = await _roleRepository.UpdateRoleAsync(existRole);
+
+                return new ApiResponse<Role>("success", "Cập nhật quyền thành công", updateRole);
+            }
+            catch (Exception ex)
+            {
+                return new ApiResponse<Role>("error", "Lỗi cập nhật quyền", null);
+            }
+            
         }
 
-        public async Task DeleteRoleAsync(int id)
+        public async Task<ApiResponse<Role>> DeleteRoleAsync(int id)
         {
-            await _roleRepository.DeleteRoleAsync(id);
+            try
+            {
+                var existRole = await _roleRepository.GetRoleByIdAsync(id);
+                if (existRole == null)
+                {
+                    return new ApiResponse<Role>("error", "Không tìm thấy quyền", null);
+                }
+                await _roleRepository.DeleteRoleAsync(id);
+                return new ApiResponse<Role>("success", "Xóa quyền thành công", null);
+
+            }
+            catch (Exception ex)
+            {
+                return new ApiResponse<Role>("error", "Lỗi xóa quyền", null);
+            }
+            
         }
     }
 }
